web/backend/pyproject.toml
[project]
name = "korean-voice-tutor-web"
version = "2.0.0"
description = "Korean Voice Tutor - Web Version (FastAPI + PTT)"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0,<14.0",
    "python-dotenv>=1.0.0",
    "openai>=1.102.0",
    "pydantic>=2.0.0",
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",  # For testing FastAPI
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
